feat: 카카오 로그인 추가

+++ src/main/java/DMU/demo/chat/domain/entity/ChatMessage.java
@@ -25,10 +25,10 @@ public class ChatMessage {
     private String roomId;
 
     @Column(nullable = false)
-    private int userId;
+    private long userId;
 
     @Column(nullable = false)
-    private int sender;
+    private long sender;
 
     @Column(nullable = false)
     private String message;
